refactor: adjust folders and files
