+++ src/pages/sapatos/sandalias/components/typesSandalsCards/index.tsx
