+++ src/pages/sapatos/sandalias/index.tsx
@@ -1,5 +1,5 @@
 import { BreadCrumb } from './components/breadCrumb';
-import { TypesSandalsCards } from './components/TypesSandalsCards';
+import { TypesSandalsCards } from './components/typesSandalsCards';
 
 export default function SandalsPDC() {
   return (
